subcommand for positioning opmode
