+++ rinex-cli/src/cli/mod.rs
@@ -1,5 +1,5 @@
 use clap::{value_parser, Arg, ArgAction, ArgMatches, ColorChoice, Command};
-use log::{info};
+use log::info;
 use map_3d::{ecef2geodetic, geodetic2ecef, Ellipsoid};
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
@@ -22,8 +22,6 @@ mod time_binning;
 mod qc;
 // positioning mode
 mod positioning;
-
-
 
 pub struct Cli {
     /// Arguments passed by user
@@ -59,7 +57,7 @@ impl Context {
      * Utility to determine the most major filename stem,
      * to be used as the session workspace
      */
-    fn context_stem(data: &RnxContext) -> String {
+    pub fn context_stem(data: &RnxContext) -> String {
         let ctx_major_stem: &str = data
             .rinex_path()
             .expect("failed to determine a context name")
