+++ rinex-cli/src/cli/time_binning.rs
@@ -6,13 +6,13 @@ pub fn subcommand() -> Command {
     Command::new("tbin")
         .long_flag("tbin")
         .arg_required_else_help(true)
-        .about("Time binning: time reframing. Subdvide RINEX files into a batch of a specific duration each.")
+        .about("Time binning. Split RINEX files into a batch of equal duration.")
         .arg(
             Arg::new("interval")
                 .value_parser(value_parser!(Duration))
                 .value_name("Duration")
                 .action(ArgAction::Set)
                 .required(true)
-                .help("Duration")
+                .help("Duration"),
         )
 }
